Spring-7. Цвета
