+++ app/build.gradle
@@ -11,7 +11,7 @@ android {
     defaultConfig {
         applicationId "com.yakushev.spring"
         minSdk 24
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -41,18 +41,17 @@ android {
     composeOptions {
         kotlinCompilerExtensionVersion '1.4.4'
     }
-    packagingOptions {
-        resources {
-            excludes += '/META-INF/{AL2.0,LGPL2.1}'
-        }
-    }
+//    packagingOptions {
+//        resources {
+//            excludes += '/META-INF/{AL2.0,LGPL2.1}'
+//        }
+//    }
 }
 
 dependencies {
     implementation 'androidx.core:core-ktx:1.10.0'
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.6.1'
     implementation 'androidx.activity:activity-compose:1.7.1'
-    implementation "androidx.fragment:fragment-ktx:1.5.7"
 
     //compose
     implementation "androidx.compose.ui:ui:$compose_version"
@@ -65,12 +64,6 @@ dependencies {
     //dagger
     implementation "com.google.dagger:dagger:$dagger_version"
     kapt "com.google.dagger:dagger-compiler:$dagger_version"
-
-    //test
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
-    androidTestImplementation "androidx.compose.ui:ui-test-junit4:$compose_version"
 
     //debug
     debugImplementation "androidx.compose.ui:ui-tooling:$compose_version"
