+++ app/src/main/java/com/yakushev/spring/presentation/MainActivity.kt
@@ -3,9 +3,11 @@ package com.yakushev.spring.presentation
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.core.view.WindowCompat
+import androidx.core.view.WindowInsetsCompat
+import androidx.core.view.WindowInsetsControllerCompat
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.compose.rememberNavController
-import com.yakushev.spring.navigation.Route
 import com.yakushev.spring.navigation.SetupNavHost
 import com.yakushev.spring.ui.theme.SpringTheme
 import javax.inject.Inject
@@ -32,6 +34,14 @@ class MainActivity : ComponentActivity() {
                     exitClick = ::finishAffinity
                 )
             }
+        }
+    }
+
+    private fun hideSystemUI() {
+        WindowCompat.setDecorFitsSystemWindows(window, false)
+        WindowInsetsControllerCompat(window, window.decorView).let { controller ->
+            controller.hide(WindowInsetsCompat.Type.systemBars())
+            controller.systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
         }
     }
 }
